feat: use talloc in userlib
1. add slint-helper to simplify the gui display

+++ userlib/src/memory.rs
@@ -5,11 +5,11 @@ pub fn brk(addr: usize) -> isize {
 }
 
 pub fn sbrk(incr: isize) -> isize {
-    let addr = brk(0) as isize;
+    let addr = brk(0);
     if incr == 0 {
         return addr;
     }
-    let new_addr = brk((addr + incr) as usize) as isize;
+    let new_addr = brk((addr + incr) as usize);
     if new_addr == addr + incr {
         return addr;
     }
